Added driver code to allocate a major number and minor number for a device
In this example the major and minor number will be reserved and the user
is responsible to create the device file using mknode command. In the next
example, we will create the device file from the driver code itself. In this
example, after the device file is created using the mknode command and the
user do the file operation on the device file which is created using the mknode,
the read/write/open.. functions will be called (Again, those read/write/open..
functions are not implemented in this driver and this will also be covered
in the next example)