Added kernel driver code for GPIO driver
This driver will be checking, requesting a GPIO and toggle the value
on and off in a seperate thread. The user can set the dealy of toggling
through the write operation. While removing the driver from the OS the
thread will be stopped and port will be released.
The mapping of physical GPIO number and actual kernel number are different.
The actual kernel number can be seen from `cat /sys/kernel/debug/gpio`.
For raspberrypi-4B the kernel number is starting from 512.